Approach #3 to indexing: xy->ij->qr
Refined the qr-col/row approach for storing the map in the memory.

Now, we can have upto 3 different indexing:

xy: dependent on the centering and origin
ij: centering and origin independent indexing (row/col with the actual
basis vectors (NE & S), with the origin being top-left
qr: 1:1 mapping to a matrix so that no null cells are left (e.g., for a
6x7 'rectangle' instead of allocating 8x7 = 56 cells, it is mapped unto
this 6x7=42 cells matrix).

This distinction will be (/have already been) explained in the wiki.